requirements.txt
fastapi
uvicorn
opencv-python
numpy
shapely
pillow

python-multipart
pillow
numpy-stl
scikit-image
python-stl
aiofiles